feat(musicservice): allow android apps to keep media notification when playback is stopped or failed
Some apps are playing internet stream directly. F.e online radio apps or so. In
such cases pause doesn't make any sense since the buffer size is limited. In such
cases most apps are uisng Stop instead of the Pause. Menwhile users need a
convinient way to start the steram again. The feature allows to specify a list of
states when app wants to keep the media session notification active. It doesn't
mean that the app will keep the foreground service running, so it can be killed
by the system at any time.